backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0

# Database Drivers
neo4j==5.25.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
asyncpg==0.30.0
databases==0.9.0

# Vector Database
qdrant-client==1.12.1

# Task Queue
celery==5.4.0
redis==5.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bcrypt==4.2.1

# API Clients
httpx==0.27.2
tweepy==4.14.0
openai==1.55.3

# Embeddings & ML
cohere==5.11.4
sentence-transformers==3.3.1
transformers==4.46.3
torch==2.5.1
ragatouille==0.0.8.post2

# WebSocket Support
python-socketio==5.11.4
websockets==14.1

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2
